# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ActivityCategory = typeof ActivityCategory[keyof typeof ActivityCategory];


export const ActivityCategory = {
  transport: 'transport',
  food: 'food',
  energy: 'energy',
  shopping: 'shopping',
} as const;

export interface Activity {
  id: number;
  sessionId: string;
  category: ActivityCategory;
  activityType: string;
  activityLabel: string;
  value: number;
  unit: string;
  co2Amount: number;
  loggedAt: string;
}

export type ActivityInputCategory = typeof ActivityInputCategory[keyof typeof ActivityInputCategory];


export const ActivityInputCategory = {
  transport: 'transport',
  food: 'food',
  energy: 'energy',
  shopping: 'shopping',
} as const;

export interface ActivityInput {
  sessionId: string;
  category: ActivityInputCategory;
  activityType: string;
  value: number;
}

export interface CategoryBreakdown {
  category: string;
  co2Amount: number;
  percentage: number;
}

export interface DailyData {
  date: string;
  co2Amount: number;
}

export interface ActivitySummary {
  totalCo2: number;
  byCategory: CategoryBreakdown[];
  dailyAverage: number;
  globalAverage: number;
  treeEquivalent: number;
  flightHoursEquivalent: number;
  weeklyData: DailyData[];
}

export type ChallengeCategory = typeof ChallengeCategory[keyof typeof ChallengeCategory];


export const ChallengeCategory = {
  transport: 'transport',
  food: 'food',
  energy: 'energy',
  shopping: 'shopping',
} as const;

export type ChallengeDifficulty = typeof ChallengeDifficulty[keyof typeof ChallengeDifficulty];


export const ChallengeDifficulty = {
  easy: 'easy',
  medium: 'medium',
  hard: 'hard',
} as const;

export interface Challenge {
  id: number;
  title: string;
  description: string;
  category: ChallengeCategory;
  co2Reduction: number;
  icon: string;
  difficulty: ChallengeDifficulty;
}

export interface ChallengeCompletion {
  id: number;
  sessionId: string;
  challengeId: number;
  completedAt: string;
}

export interface ChallengeCompletionInput {
  sessionId: string;
}

export interface LeaderboardEntry {
  rank: number;
  displayName: string;
  co2Reduced: number;
  challengesCompleted: number;
  totalCo2Logged: number;
  topCategory: string;
  isCurrentUser: boolean;
}

export interface CoachAdviceInput {
  sessionId: string;
}

export type CoachTipEffort = typeof CoachTipEffort[keyof typeof CoachTipEffort];


export const CoachTipEffort = {
  low: 'low',
  medium: 'medium',
  high: 'high',
} as const;

export interface CoachTip {
  category: string;
  tip: string;
  impact: string;
  effort: CoachTipEffort;
}

export interface CoachAdvice {
  message: string;
  tips: CoachTip[];
  focusArea: string;
  weeklyGoal: string;
}

export type ListActivitiesParams = {
sessionId: string;
days?: number;
};

export type GetActivitySummaryParams = {
sessionId: string;
days?: number;
};

export type ListChallengeCompletionsParams = {
sessionId: string;
};

export type GetLeaderboardParams = {
sessionId?: string;
};