# lib/api-zod/src/generated/types/activityInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityInputCategory } from './activityInputCategory';

export interface ActivityInput {
  sessionId: string;
  category: ActivityInputCategory;
  activityType: string;
  value: number;
}