# lib/api-zod/src/generated/types/challenge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ChallengeCategory } from './challengeCategory';
import type { ChallengeDifficulty } from './challengeDifficulty';

export interface Challenge {
  id: number;
  title: string;
  description: string;
  category: ChallengeCategory;
  co2Reduction: number;
  icon: string;
  difficulty: ChallengeDifficulty;
}