# lib/api-zod/src/generated/types/challengeCompletion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ChallengeCompletion {
  id: number;
  sessionId: string;
  challengeId: number;
  completedAt: string;
}