# lib/api-zod/src/generated/types/coachTip.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CoachTipEffort } from './coachTipEffort';

export interface CoachTip {
  category: string;
  tip: string;
  impact: string;
  effort: CoachTipEffort;
}