# lib/api-zod/src/generated/types/dailyData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EcoTrace Carbon Footprint Awareness Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface DailyData {
  date: string;
  co2Amount: number;
}